Refactor ClinicCreateFlow
