--- app/common/containers/blocks/CreateClinicStep3/styles.scss
@@ -1,45 +1,0 @@
-.section {
-  display: block;
-
-  &__title {
-    margin-bottom: 29px;
-  }
-
-  &__form {
-    &__title {
-      margin-bottom: 30px;
-      color: #3c4858;
-      font-size: 18px;
-      font-weight: 700;
-    }
-  }
-}
-
-.options {
-  display: block;
-  position: relative;
-
-  &__new {
-    margin-top: 25px;
-    vertical-align: middle;
-
-    a {
-      cursor: pointer;
-    }
-
-    img {
-      width: 27px;
-      height: 27px;
-      margin-right: 15px;
-      margin-top: -2px;
-    }
-
-    span {
-      line-height: 27px;
-      vertical-align: top;
-      color: #3c4858;
-      font-weight: 700;
-      text-transform: uppercase;
-    }
-  }
-}
